pyproject.toml
[tool.poetry]
name = "scandi_dpr"
version = "0.0.0"
description = "Dense Passage Retrieval models for the Scandinavian languages."
authors = [
    "Dan Saattrup Nielsen <dan.nielsen@alexandra.dk>",
]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
hydra-core = "^1.3.2"
transformers = "^4.33.2"
torch = "^2.0.1"
datasets = "^2.14.5"
sentencepiece = ">=0.1.99,<1.0.0"
accelerate = ">=0.23.0,<1.0.0"
evaluate = ">=0.4.0,<1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
pytest-xdist = "^3.2.1"
pytest-cov = "^4.0.0"
pre-commit = "^3.3.1"
pdoc = "^13.1.1"
black = "^23.3.0"
isort = "^5.12.0"
readme-coverage-badger = ">=0.1.2,<1.0.0"

[tool.pytest.ini_options]
minversion = "7.0"
addopts = [
    '--verbose',
    '--durations=10',
    '--color=yes',
    '-s',
    '-vv',
    '--doctest-modules',
    '--cov=src/scandi_dpr',
    '-n 2',
]
xfail_strict = true
filterwarnings = ["ignore"]
log_cli_level = "info"
testpaths = ["tests", "src/scandi_dpr"]

[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
/(
	\.git
| \.hg
| \.mypy_cache
| \.tox
| \.venv
| _build
| buck-out
| build
)/
'''

[tool.ruff]
target-version = "py311"

[tool.ruff.extend-per-file-ignores]
"__init__.py" = ["F401"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
